Cast to int before printing with int format
Original-revision: 643

+++ mount.ocfs2/mount.ocfs2.c
@@ -807,7 +807,7 @@ int main(int argc, char **argv)
 
 	ret = add_me_to_group(hbuuid, mo.dev);
 	if (ret < 0) {
-		fprintf(stderr, "%s: Error '%d' while adding to group\n", progname, ret);
+		fprintf(stderr, "%s: Error '%d' while adding to group\n", progname, (int)ret);
 		goto bail;
 	}
